Revert "add middleware"
