+++ helpers/index.js
@@ -1,4 +1,3 @@
 const HttpError = require("./HttpError");
-const ctrlWrapper = require("./ctrlWrapper");
 
-module.exports = { HttpError, ctrlWrapper };
+module.exports = { HttpError };
