--- middlewares/index.js
@@ -1,3 +1,0 @@
-const validateBody = require("./validateBody");
-
-module.exports = { validateBody };
