--- middlewares/validateBody.js
@@ -1,14 +1,0 @@
-const { HttpError } = require("../helpers");
-
-const validateBody = (schema) => {
-  const fn = (req, res, next) => {
-    const { error } = schema.validate(req.body);
-    if (error) {
-      next(HttpError(400, error.message));
-    }
-    next();
-  };
-  return fn;
-};
-
-module.exports = validateBody;
